backend/auth/requirements.txt
deta
fastapi
uvicorn[standard]
email-validator
passlib[bcrypt]
python-jose[cryptography]
gunicorn
flower
requests